Moved ekg stuff into its own module
