+++ retcon.cabal
@@ -32,6 +32,7 @@ library
     Retcon.Identifier
     Retcon.Store
     Retcon.Monad
+    Retcon.Network.Ekg
     Retcon.Network.Protocol
     Retcon.Network.Server
     Retcon.Network.Client
